requirements.txt
bcrypt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
gunicorn
psycopg2
flask_bootstrap
flask_Datepicker